+++ src/main/java/com/nhom7/coworkingspace/service/BookingService.java
@@ -1,5 +1,6 @@
 package com.nhom7.coworkingspace.service;
 
+import com.nhom7.coworkingspace.dto.request.BookingHistoryRequest;
 import com.nhom7.coworkingspace.dto.request.BookingRequest;
 import com.nhom7.coworkingspace.dto.request.BookingSearchRequest;
 import com.nhom7.coworkingspace.dto.response.BookingResponse;
@@ -11,6 +12,8 @@ public interface BookingService {
     Booking changeStatus(Long bookingId, String newStatus);
 
     BookingResponse createBooking(BookingRequest request, String userEmail);
+
+    PageResponse<BookingResponse> getMyBookingHistory(String userEmail, BookingHistoryRequest request);
 
     BookingResponse cancelBooking(Long bookingId, String userEmail);